Increasing the maximum file limit
By default, the HDF4 libraries have a fixed limit on the number
of HDF4 simultaneously opened (it's 32). This limit is too small
for some applications, and the value can be increased safely to
much higher figures, which is what this patch does. The patch
is provided, and also referred to in the `meta.yaml` file. This
patch had been accepted in the previous conda recipes github
repository, but it's not visible on the conda-forge feedstock.